Selenium grid setup done

+++ src/test/java/testBase/BaseClass.java
@@ -3,6 +3,7 @@ package testBase;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URL;
 import java.sql.Date;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
@@ -13,10 +14,14 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;//log4j
 import org.apache.logging.log4j.Logger;   //log4j
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
@@ -43,14 +48,58 @@ public class BaseClass {
 		//loading log4j file
 		logger=LogManager.getLogger(this.getClass());//Log4j
 		
-		//launching browser based on condition
-		switch(br.toLowerCase())
+			
+		// for selenium grid remote envirment launching
+		if(p.getProperty("execution_env").equalsIgnoreCase("remote"))
 		{
-		case "chrome": driver=new ChromeDriver(); break;
-		case "edge": driver=new EdgeDriver(); break;
-		default: System.out.println("No matching browser..");
-					return;
+			DesiredCapabilities capabilities=new DesiredCapabilities();
+			
+			//os
+			if(os.equalsIgnoreCase("windows"))
+			{
+				capabilities.setPlatform(Platform.WIN11);
+			}
+			else if (os.equalsIgnoreCase("mac"))
+			{
+				capabilities.setPlatform(Platform.MAC);
+			}
+			else
+			{
+				System.out.println("No matching os");
+				return;
+			}
+
+			//browser
+			switch(br.toLowerCase())
+			{
+			case "chrome": capabilities.setBrowserName("chrome"); break;
+			case "edge": capabilities.setBrowserName("MicrosoftEdge"); break;
+			default: System.out.println("No matching browser"); return;
+			}
+			
+			driver=new RemoteWebDriver(new URL("http://192.168.150.250:4444/wd/hub"),capabilities);
 		}
+		
+				
+		if(p.getProperty("execution_env").equalsIgnoreCase("local"))
+		{
+
+			switch(br.toLowerCase())
+			{
+			case "chrome" : driver=new ChromeDriver(); break;
+			case "edge" : driver=new EdgeDriver(); break;
+			case "firefox": driver=new FirefoxDriver(); break;
+			default : System.out.println("Invalid browser name.."); return;
+			}
+		}
+		
+			
+		/*
+		 * //launching browser based on condition on local switch(br.toLowerCase()) {
+		 * case "chrome": driver=new ChromeDriver(); break; case "edge": driver=new
+		 * EdgeDriver(); break; default: System.out.println("No matching browser..");
+		 * return; }
+		 */
 		
 		driver.manage().deleteAllCookies();
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
